More work on calculate response.
